+++ src/components/ui/components/BreadcrumbUi/config.js
@@ -1,0 +1,9 @@
+const config = {
+  name: 'BreadcrumbUi',
+  label: 'Breadcrumb',
+  icon: 'angle-right',
+  group: 'components',
+  listable: true,
+}
+
+export default config