+++ src/components/ui/components/ModalUi/config.js
@@ -1,0 +1,9 @@
+const config = {
+  name: 'ModalUi',
+  label: 'Modal',
+  icon: 'angle-right',
+  group: 'components',
+  listable: false,
+}
+
+export default config