+++ src/components/ui/elements/ButtonUi/config.js
@@ -1,0 +1,9 @@
+const config = {
+  name: 'ButtonUI',
+  label: 'Button',
+  icon: 'angle-right',
+  group: 'elements',
+  listable: false,
+}
+
+export default config