+++ src/components/ui/elements/ContentUi/config.js
@@ -1,0 +1,9 @@
+const config = {
+  name: 'ContentUi',
+  label: 'Content',
+  icon: 'angle-right',
+  group: 'elements',
+  listable: true,
+}
+
+export default config