+++ src/components/ui/elements/DeleteUi/config.js
@@ -1,0 +1,9 @@
+const config = {
+  name: 'DeleteUi',
+  label: 'Delete',
+  icon: 'angle-right',
+  group: 'elements',
+  listable: false,
+}
+
+export default config