+++ src/components/ui/elements/NotificationUi/NotificationUi.vue
@@ -1,0 +1,21 @@
+<template lang="pug">
+  .NotificationUi.notification
+    button.delete
+    |   Lorem ipsum dolor sit amet, consectetur
+    |   adipiscing elit lorem ipsum dolor.
+    strong Pellentesque risus mi
+    | , tempus quis placerat ut, porta nec nulla. Vestibulum rhoncus ac ex sit amet fringilla. Nullam gravida purus diam, et dictum
+    a felis venenatis
+    |  efficitur. Sit amet,
+    |   consectetur adipiscing elit
+</template>
+
+<script>
+export default {
+  name: 'NotificationUi',
+}
+</script>
+
+<style lang="stylus">
+// .NotificationUi
+</style>