+++ src/components/ui/elements/configs.js
@@ -1,0 +1,23 @@
+import BoxUi from '@/components/ui/elements/BoxUi/config'
+import ButtonUi from '@/components/ui/elements/ButtonUi/config'
+import ContentUi from '@/components/ui/elements/ContentUi/config'
+import DeleteUi from '@/components/ui/elements/DeleteUi/config'
+import IconUi from '@/components/ui/elements/IconUi/config'
+import ImageUi from '@/components/ui/elements/ImageUi/config'
+import NotificationUi from '@/components/ui/elements/NotificationUi/config'
+import ProgressUi from '@/components/ui/elements/ProgressUi/config'
+import TableUi from '@/components/ui/elements/TableUi/config'
+import TitleUi from '@/components/ui/elements/TitleUi/config'
+
+export {
+  BoxUi,
+  ButtonUi,
+  ContentUi,
+  DeleteUi,
+  IconUi,
+  ImageUi,
+  NotificationUi,
+  ProgressUi,
+  TableUi,
+  TitleUi
+}